[#60] Fix html anchors
For the very first rst file toc entry, Sphinx doesn't add an anchor,
but instead uses file's title.
Because of that, some links in Table of contents of PDF document
are not correct, resulting in wrong page numbers.
Fix replaces all such anchors with the correct ones, after the
creation of HTML document.